website/requirement.txt
asgiref==3.2.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.4
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==1.1.0
django-filters==0.2.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-rest-knox==4.1.0
django-taggit==1.2.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
gunicorn==20.0.4
idna==2.9
Markdown==3.2.1
markdown2==2.3.8
Pillow==7.0.0
pkg-resources==0.0.0
pycparser==2.20
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
